+++ cmd/forklift/main.go
@@ -113,9 +113,9 @@ const (
 	bundleMinVersion = "v0.7.0"
 	// newBundleVersion is the Forklift version reported in new staged pallet bundles made by Forklift.
 	// Older versions of the Forklift tool cannot use such bundles.
-	newBundleVersion = "v0.8.0-alpha.2"
+	newBundleVersion = "v0.8.0-alpha.6"
 	// newStageStoreVersion is the Forklift version reported in a stage store initialized by Forklift.
-	// Older versions of the Forklift tool cannot use the state store.
+	// Older versions of the Forklift tool cannot use the stage store.
 	newStageStoreVersion = "v0.7.0"
 	// fallbackVersion is the version reported which the Forklift tool reports itself as if its actual
 	// version is unknown.